requirements.txt
aniso8601==1.2.0
asn1crypto==0.22.0
beautifulsoup4==4.6.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==1.9
cssselect==1.0.1
Django==1.11.2
django-cors-headers==2.1.0
djangorestframework==3.6.3
Flask==0.12.1
Flask-Ask==0.9.4
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
lxml==3.8.0
MarkupSafe==1.0
pycparser==2.17
Pygments==2.2.0
pyOpenSSL==17.0.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.18.1
six==1.10.0
urllib3==1.21.1
Werkzeug==0.12.2

